docs: made changes to readme
readme was a bit parse, and lacked an overall style guide, as
per the last review for this project. I added some information regarding
cloning or downloading the source code, with ordered steps for easy
implementation.

The code should be easy enough to read and/or download at this point.
One's favorite code editor should be used, and Git Bash is recommended
for all git processes.

Resolves most recent review fixes